+++ src/Config/config.php
@@ -21,6 +21,7 @@ return [
                     'login' => 'auth/classic/login',
                     'register' => 'auth/classic/register',
                     'me' => 'auth/classic/me',
+                    'logout' => 'auth/classic/logout'
                 ],
             ],
         ]