+++ src/Routes/api.php
@@ -11,6 +11,7 @@ Route::prefix(config('waccount.routes.prefix'))->group(function () {
             Route::post('login', 'login')->name('login');
             Route::post('register', 'register')->name('register');
             Route::post('me', 'me')->name('me');
+            Route::post('logout', 'logout')->name('logout');
         });
     });